Update polymarket extension (#19089)
* Update polymarket extension

- fix: owner overwrite
- chore: fix lint issues
- chore: add publish script
- fix: remove incorrect import
- chore: fix spelling
- refactor: move market list item into dedicated function
- chore: update descriptions and add screenshots
- refactor: move shared logic to common
- chore: move urls to constants file
- feat: search currently active markets
- chore: move types & utils into dedicated filesOD
- Initial commit

* fix: pr comments

- Update Changelog format
- Fixed incorrect parameter destructuring
- Use `.toString()`
- Add subtitle

* fix: use market slug as key

* chore: run prettier

* chore: remove .fleet

* Update CHANGELOG.md and optimise images

---------
