+++ extensions/polymarket/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Polymarket Changelog
 
+## [PolyMarket Search] - 2025-05-14
+- Added ability to search PolyMarket for a specific market.
+
 ## [Remove old election markets] - 2024-11-18
 Removed 2024 US Election Odds and Swing States
