+++ extensions/polymarket/package.json
@@ -2,9 +2,12 @@
   "$schema": "https://www.raycast.com/schemas/extension.json",
   "name": "polymarket",
   "title": "Polymarket",
-  "description": "Search and view top markets on Polymarket",
+  "description": "Search and view markets on Polymarket",
   "icon": "polymarket.png",
   "author": "Osprey",
+  "contributors": [
+    "dkloecker"
+  ],
   "categories": [
     "News",
     "Finance"
@@ -16,10 +19,18 @@
       "title": "Top Markets",
       "description": "Shows top markets by 24h volume",
       "mode": "view"
+    },
+    {
+      "name": "search-for-market",
+      "title": "Polymarket Search",
+      "subtitle": "Search Active Markets",
+      "description": "Search Polymarket for specific market",
+      "mode": "view"
     }
   ],
   "dependencies": {
     "@raycast/api": "^1.84.6",
+    "@raycast/utils": "^1.19.1",
     "node-fetch": "^3.3.2"
   },
   "devDependencies": {
@@ -31,6 +42,7 @@
     "typescript": "^5.4.5"
   },
   "scripts": {
+    "publish": "npx @raycast/api@latest publish",
     "build": "ray build -e dist",
     "dev": "ray develop",
     "fix-lint": "ray lint --fix",
