+++ extensions/polymarket/src/constants.ts
@@ -1,0 +1,4 @@
+const POLY_GAMMA_URL = "https://gamma-api.polymarket.com/";
+const POLY_REST_URL = "https://polymarket.com/api/";
+
+export { POLY_GAMMA_URL, POLY_REST_URL };
